added verbose logging

+++ src/entities/write.ts
@@ -1,7 +1,11 @@
-import { writeFile } from 'fs-extra';
+import { pathExistsSync, writeFile } from 'fs-extra';
 import { join } from 'path';
 
-export const writeEntity = async (name: string, directory: string, content: string): Promise<void> => {
+export const writeEntity = async (name: string, directory: string, content: string): Promise<string> => {
   const path = join(directory, `${name}.entity.ts`);
+
+  if (pathExistsSync(path)) throw new Error('EXISTS');
   await writeFile(path, content);
+
+  return path;
 };
